Override rack logger, change file structure
